@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 
 typedef Widget ItemBuilder<T>(BuildContext context, T entry, int index);
-typedef Future<List<T>> PageFuture<T>(int pageIndex);
-typedef Widget ErrorBuilder(BuildContext context, Object error);
+typedef Future<List<T>> PageFuture<T>(int? pageIndex);
+typedef Widget ErrorBuilder(BuildContext context, Object? error);
 typedef Widget NoItemsFoundBuilder(BuildContext context);
 typedef Widget RetryBuilder(BuildContext context, RetryCallback retryCallback);
 typedef void RetryCallback();
@@ -24,14 +24,14 @@ typedef Widget PagewiseBuilder<T>(PagewiseState<T> state);
 ///  * [PagewiseListView], a [Pagewise] implementation of [ListView](https://docs.flutter.io/flutter/widgets/ListView-class.html)
 abstract class Pagewise<T> extends StatefulWidget {
   /// The number  of entries per page
-  final int pageSize;
+  final int? pageSize;
 
   /// Called whenever a new page (or batch) is to be fetched
   ///
   /// It is provided with the page index, and expected to return a [Future](https://api.dartlang.org/stable/1.24.3/dart-async/Future-class.html) that
   /// resolves to a list of entries. Please make sure to return only [pageSize]
   /// or less entries (in the case of the last page) for each page.
-  final PageFuture<T> pageFuture;
+  final PageFuture<T>? pageFuture;
 
   /// Called with an error object if an error occurs when loading the page
   ///
@@ -43,7 +43,7 @@ abstract class Pagewise<T> extends StatefulWidget {
   /// }
   /// ```
   /// If not specified, a [Text] containing the error will be displayed
-  final ErrorBuilder errorBuilder;
+  final ErrorBuilder? errorBuilder;
 
   /// Called when no items are found
   ///
@@ -55,7 +55,7 @@ abstract class Pagewise<T> extends StatefulWidget {
   ///    return Text('No Items Found!');
   ///  }
   ///  ```
-  final NoItemsFoundBuilder noItemsFoundBuilder;
+  final NoItemsFoundBuilder? noItemsFoundBuilder;
 
   /// Called to build each entry in the view.
   ///
@@ -96,7 +96,7 @@ abstract class Pagewise<T> extends StatefulWidget {
   /// You don't have to provide this parameter unless you want to control or
   /// listen to the data that Pagewise fetches. Review the documentation of
   /// [PagewiseLoadController] for more details
-  final PagewiseLoadController<T> pageLoadController;
+  final PagewiseLoadController<T>? pageLoadController;
 
   /// Creates a pagewise widget.
   ///
@@ -105,12 +105,12 @@ abstract class Pagewise<T> extends StatefulWidget {
   Pagewise(
       {this.pageSize,
       this.pageFuture,
-      Key key,
+      Key? key,
       this.pageLoadController,
       this.noItemsFoundBuilder,
-      @required this.itemBuilder,
+      required this.itemBuilder,
       this.errorBuilder,
-      @required this.builder})
+      required this.builder})
       : assert((pageLoadController == null &&
                 pageSize != null &&
                 pageFuture != null) ||
@@ -123,102 +123,102 @@ abstract class Pagewise<T> extends StatefulWidget {
   PagewiseState<T> createState() => PagewiseState<T>();
 }
 
-class PagewiseState<T> extends State<Pagewise<T>> {
-  PagewiseLoadController<T> _controller;
+class PagewiseState<T> extends State<Pagewise<T?>> {
+  PagewiseLoadController<T?>? _controller;
 
-  PagewiseLoadController<T> get _effectiveController =>
+  PagewiseLoadController<T?>? get _effectiveController =>
       widget.pageLoadController ?? this._controller;
 
-  VoidCallback _controllerListener;
+  late VoidCallback _controllerListener;
 
   @override
   void initState() {
     super.initState();
 
     if (widget.pageLoadController == null) {
-      this._controller = PagewiseLoadController<T>(
+      this._controller = PagewiseLoadController<T?>(
           pageFuture: widget.pageFuture, pageSize: widget.pageSize);
     }
 
-    this._effectiveController.init();
+    this._effectiveController!.init();
 
     this._controllerListener = () {
       setState(() {});
     };
 
-    this._effectiveController.addListener(this._controllerListener);
+    this._effectiveController!.addListener(this._controllerListener);
   }
 
   @override
   void dispose() {
     super.dispose();
-    this._effectiveController.removeListener(this._controllerListener);
+    this._effectiveController!.removeListener(this._controllerListener);
   }
 
   @override
   void didUpdateWidget(Pagewise oldWidget) {
-    super.didUpdateWidget(oldWidget);
+    super.didUpdateWidget(oldWidget as Pagewise<T>);
     if (widget.pageLoadController == null &&
         oldWidget.pageLoadController != null) {
-      oldWidget.pageLoadController.removeListener(this._controllerListener);
+      oldWidget.pageLoadController!.removeListener(this._controllerListener);
       this._controller = PagewiseLoadController<T>(
-          pageFuture: oldWidget.pageLoadController.pageFuture,
-          pageSize: oldWidget.pageLoadController.pageSize);
-      this._effectiveController.addListener(this._controllerListener);
-      this._effectiveController.init();
+          pageFuture: oldWidget.pageLoadController!.pageFuture,
+          pageSize: oldWidget.pageLoadController!.pageSize);
+      this._effectiveController!.addListener(this._controllerListener);
+      this._effectiveController!.init();
     } else if (widget.pageLoadController != null &&
         oldWidget.pageLoadController == null) {
-      this._controller.removeListener(this._controllerListener);
+      this._controller!.removeListener(this._controllerListener);
       this._controller = null;
-      this._effectiveController.addListener(this._controllerListener);
-      this._effectiveController.init();
+      this._effectiveController!.addListener(this._controllerListener);
+      this._effectiveController!.init();
     } else if (widget.pageLoadController != null &&
         (widget.pageLoadController != oldWidget.pageLoadController)) {
-      oldWidget.pageLoadController.removeListener(this._controllerListener);
-      this._effectiveController.addListener(this._controllerListener);
-      this._effectiveController.init();
+      oldWidget.pageLoadController!.removeListener(this._controllerListener);
+      this._effectiveController!.addListener(this._controllerListener);
+      this._effectiveController!.init();
     }
   }
 
-  int get _itemCount => this._effectiveController.loadedItems.length + 1;
+  int get _itemCount => this._effectiveController!.loadedItems!.length + 1;
 
   @override
   Widget build(BuildContext context) {
     return widget.builder(this);
   }
 
-  Widget _itemBuilder(BuildContext context, int index) {
+  Widget? _itemBuilder(BuildContext context, int index) {
     // The total number of widgets, is the number of loaded items
     // plus 1 for the loader
-    final total = this._effectiveController.loadedItems.length + 1;
+    final total = this._effectiveController!.loadedItems!.length + 1;
 
     if (index >= total) {
       return null;
     }
     if (index == total - 1) {
-      if (this._effectiveController.noItemsFound) {
+      if (this._effectiveController!.noItemsFound) {
         return this._getNoItemsFoundWidget();
       }
 
-      if (this._effectiveController.error != null) {
-        return this._getErrorWidget(this._effectiveController.error);
+      if (this._effectiveController!.error != null) {
+        return this._getErrorWidget(this._effectiveController!.error);
       }
 
-      if (this._effectiveController.hasMoreItems) {
-        this._effectiveController.fetchNewPage();
+      if (this._effectiveController!.hasMoreItems!) {
+        this._effectiveController!.fetchNewPage();
         return this._getLoadingWidget();
       } else {
         return Container();
       }
     } else {
-      if (index >= this._effectiveController.loadedItems.length) {
+      if (index >= this._effectiveController!.loadedItems!.length) {
         // this means that the function is asking for an element from the
         // appended items, so we return an empty container
         return Container();
       }
       // Otherwise, we return the actual item
       return widget.itemBuilder(
-          context, this._effectiveController.loadedItems[index], index);
+          context, this._effectiveController!.loadedItems![index], index);
     }
   }
 
@@ -229,21 +229,21 @@ class PagewiseState<T> extends State<Pagewise<T>> {
   Widget _getNoItemsFoundWidget() {
     return this._getStandardContainer(
         child: widget.noItemsFoundBuilder != null
-            ? widget.noItemsFoundBuilder(context)
+            ? widget.noItemsFoundBuilder!(context)
             : Container());
   }
 
-  Widget _getErrorWidget(Object error) {
+  Widget _getErrorWidget(Object? error) {
     return this._getStandardContainer(
         child: widget.errorBuilder != null
-            ? widget.errorBuilder(context, this._effectiveController.error)
+            ? widget.errorBuilder!(context, this._effectiveController!.error)
             : Text('Error: $error',
                 style: TextStyle(
                     color: Theme.of(context).disabledColor,
                     fontStyle: FontStyle.italic)));
   }
 
-  Widget _getStandardContainer({Widget child}) {
+  Widget _getStandardContainer({Widget? child}) {
     return Padding(
         padding: const EdgeInsets.symmetric(vertical: 8.0),
         child: Align(
@@ -320,42 +320,42 @@ class PagewiseState<T> extends State<Pagewise<T>> {
 /// }
 /// ```
 class PagewiseLoadController<T> extends ChangeNotifier {
-  List<T> _loadedItems;
-  int _numberOfLoadedPages;
-  bool _hasMoreItems;
-  Object _error;
-  bool _isFetching;
+  List<T>? _loadedItems;
+  int? _numberOfLoadedPages;
+  bool? _hasMoreItems;
+  Object? _error;
+  late bool _isFetching;
 
   /// Called whenever a new page (or batch) is to be fetched
   ///
   /// It is provided with the page index, and expected to return a [Future](https://api.dartlang.org/stable/1.24.3/dart-async/Future-class.html) that
   /// resolves to a list of entries. Please make sure to return only [pageSize]
   /// or less entries (in the case of the last page) for each page.
-  final PageFuture<T> pageFuture;
+  final PageFuture<T>? pageFuture;
 
   /// The number  of entries per page
-  final int pageSize;
+  final int? pageSize;
 
   /// Creates a PagewiseLoadController.
   ///
   /// You must provide both the [pageFuture] and the [pageSize]
-  PagewiseLoadController({@required this.pageFuture, @required this.pageSize});
+  PagewiseLoadController({required this.pageFuture, required this.pageSize});
 
   /// The list of items that have already been loaded
-  List<T> get loadedItems => this._loadedItems;
+  List<T>? get loadedItems => this._loadedItems;
 
   /// The number of pages that have already been loaded
-  int get numberOfLoadedPages => this._numberOfLoadedPages;
+  int? get numberOfLoadedPages => this._numberOfLoadedPages;
 
   /// Whether there are still more items to load
-  bool get hasMoreItems => this._hasMoreItems;
+  bool? get hasMoreItems => this._hasMoreItems;
 
   /// The latest error that has been faced when trying to load a page
-  Object get error => this._error;
+  Object? get error => this._error;
 
   /// set to true if no data was found
   bool get noItemsFound =>
-      this._loadedItems.length == 0 && this.hasMoreItems == false;
+      this._loadedItems!.length == 0 && this.hasMoreItems == false;
 
   /// Called to initialize the controller. Same as [reset]
   void init() {
@@ -379,8 +379,10 @@ class PagewiseLoadController<T> extends ChangeNotifier {
 
       List<T> page;
       try {
-        page = await this.pageFuture(this._numberOfLoadedPages);
-        this._numberOfLoadedPages++;
+        page = await this.pageFuture!(this._numberOfLoadedPages);
+        if (this._numberOfLoadedPages != null) {
+          this._numberOfLoadedPages = this._numberOfLoadedPages! + 1;
+        }
       } catch (error) {
         this._error = error;
         this._isFetching = false;
@@ -391,7 +393,7 @@ class PagewiseLoadController<T> extends ChangeNotifier {
       // Get length accounting for possible null Future return. We'l treat a null Future as an empty return
       final int length = (page?.length ?? 0);
 
-      if (length > this.pageSize) {
+      if (length > this.pageSize!) {
         this._isFetching = false;
         throw ('Page length ($length) is greater than the maximum size (${this.pageSize})');
       }
@@ -399,7 +401,7 @@ class PagewiseLoadController<T> extends ChangeNotifier {
       if (length == 0) {
         this._hasMoreItems = false;
       } else {
-        this._loadedItems.addAll(page);
+        this._loadedItems!.addAll(page);
       }
       this._isFetching = false;
       notifyListeners();
@@ -419,27 +421,27 @@ class PagewiseListView<T> extends Pagewise<T> {
   /// All the properties are either those documented for normal [ListViews](https://docs.flutter.io/flutter/widgets/ListView-class.html),
   /// or those inherited from [Pagewise]
   PagewiseListView(
-      {Key key,
-      EdgeInsetsGeometry padding,
-      bool primary,
+      {Key? key,
+      EdgeInsetsGeometry? padding,
+      bool? primary,
       bool addSemanticIndexes = true,
-      int semanticChildCount,
+      int? semanticChildCount,
       bool shrinkWrap: false,
-      ScrollController controller,
-      PagewiseLoadController<T> pageLoadController,
-      double itemExtent,
+      ScrollController? controller,
+      PagewiseLoadController<T>? pageLoadController,
+      double? itemExtent,
       bool addAutomaticKeepAlives: true,
       Axis scrollDirection: Axis.vertical,
       bool addRepaintBoundaries: true,
-      double cacheExtent,
-      ScrollPhysics physics,
+      double? cacheExtent,
+      ScrollPhysics? physics,
       bool reverse: false,
-      int pageSize,
-      PageFuture<T> pageFuture,
-      RetryBuilder retryBuilder,
-      NoItemsFoundBuilder noItemsFoundBuilder,
-      @required ItemBuilder<T> itemBuilder,
-      ErrorBuilder errorBuilder})
+      int? pageSize,
+      PageFuture<T>? pageFuture,
+      RetryBuilder? retryBuilder,
+      NoItemsFoundBuilder? noItemsFoundBuilder,
+      required ItemBuilder<T> itemBuilder,
+      ErrorBuilder? errorBuilder})
       : super(
             pageSize: pageSize,
             pageFuture: pageFuture,
@@ -464,6 +466,7 @@ class PagewiseListView<T> extends Pagewise<T> {
                   primary: primary,
                   controller: controller,
                   itemCount: state._itemCount,
-                  itemBuilder: state._itemBuilder);
+                  itemBuilder:
+                      state._itemBuilder as Widget Function(BuildContext, int));
             });
 }
