@@ -20,7 +20,7 @@ import 'package:flutter/rendering.dart';
 typedef void OnWidgetSizeChange(Size size);
 
 class MeasureSizeRenderObject extends RenderProxyBox {
-  Size oldSize;
+  Size? oldSize;
   final OnWidgetSizeChange onChange;
 
   MeasureSizeRenderObject(this.onChange);
@@ -29,7 +29,7 @@ class MeasureSizeRenderObject extends RenderProxyBox {
   void performLayout() {
     super.performLayout();
 
-    Size newSize = child.size;
+    Size newSize = child!.size;
     if (oldSize == newSize) return;
 
     oldSize = newSize;
@@ -43,9 +43,9 @@ class MeasureSize extends SingleChildRenderObjectWidget {
   final OnWidgetSizeChange onChange;
 
   const MeasureSize({
-    Key key,
-    @required this.onChange,
-    @required Widget child,
+    Key? key,
+    required this.onChange,
+    required Widget child,
   }) : super(key: key, child: child);
 
   @override
