@@ -25,12 +25,12 @@ class ObservableStringProvider extends InheritedWidget {
   final ObservableString observable;
 
   const ObservableStringProvider(
-      {Key key, @required this.observable, @required Widget child})
+      {Key? key, required this.observable, required Widget child})
       : assert(observable != null),
         assert(child != null),
         super(key: key, child: child);
 
-  static ObservableString of(BuildContext context) => context
+  static ObservableString? of(BuildContext context) => context
       .dependOnInheritedWidgetOfExactType<ObservableStringProvider>()
       ?.observable;
 
