@@ -14,7 +14,7 @@
 
 import 'dart:html' as html;
 
-String rendererServiceAddress() {
+String? rendererServiceAddress() {
   return html
       .querySelector('meta[name=spec-renderer-service]')
       ?.getAttribute('content');
